// src/app/layout/features/pipe-line/dashboard/hash-worker.types.ts
export interface HashWorkerMessage {
  file?: string;
  timestamp: Date;
  message: string;
  thread: Thread;
  fileEventType: FileHashEvent | null; // null if not a file event
  // values computed after emission
  millisSinceLast?: number;
  percentage?: string;


}

export enum Thread {
  WORKER = 'worker',
  MAIN = 'main',
}

export enum FileHashEvent {
  SELECTED = 'selected',
  PICKED_UP = 'picked_up',
  FILE_RECEIVED = 'file_received',
  FILE_READ = 'file_read',
  HASH_COMPUTED = 'hash_computed',
  HASH_RECEIVED = 'hash_received',
  COMPLETED = "completed",
  FAIL = 'fali'
}